Trait stif_test::csif::io::SizedField

pub trait SizedField: Sized {
    const SIZE: usize;

    // Required methods
    fn csif_from(slice: &Vec<u8>, index: usize) -> Result<Self, ParsingError>;
    fn csif_into(&self) -> Vec<u8>;
}

Trait for types which can be pulled from a vector of bytes in a sized fasion, i.e., have constant size within any schema.

Required Associated Constants

const SIZE: usize

Required Methods

fn csif_from(slice: &Vec<u8>, index: usize) -> Result<Self, ParsingError>

fn csif_into(&self) -> Vec<u8>

Object Safety

This trait is not object safe.